Update become-3-season-leader.md
Hi David,
I'd like to change the way upgrade recommendations are submitted. Currently they go with emails and I have no access for example to what previous hiking chair received. I made a google form to collect future recs (same as WSC does) and I would like to have the link on the website. For now I will host the form and answers on my personal account, for the future (maybe for next year (we can talk about how to transfer the form to the club's account and maybe visualize responses on the mitoc-trips website. Thanks!
Cristian Cavedon